+++ app/build.gradle
@@ -9,7 +9,7 @@ android {
     compileSdkVersion 29
     defaultConfig {
         applicationId "io.krugosvet.dailydish.android"
-        minSdkVersion 19
+        minSdkVersion 26
         targetSdkVersion 29
         versionCode 4
         versionName "0.1.3"
@@ -59,7 +59,9 @@ android {
 
 dependencies {
     implementation project(":bindingcomponent")
-    implementation project(":db")
+    implementation project(":repository")
+    implementation project(":core")
+    implementation project(":reminder")
 
     // AndroidX Core
