+++ app/src/main/java/io/krugosvet/dailydish/android/DailyDishApplication.kt
@@ -1,8 +1,12 @@
 package io.krugosvet.dailydish.android
 
 import android.app.Application
+import io.krugosvet.android.reminder.ReminderService
+import io.krugosvet.android.reminder.injection.reminderModule
 import io.krugosvet.dailydish.android.architecture.injection.module
-import io.krugosvet.dailydish.injection.dbModule
+import io.krugosvet.dailydish.android.repository.injection.repositoryModule
+import io.krugosvet.dailydish.core.injection.coreModule
+import org.koin.android.ext.android.get
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.ext.koin.androidLogger
 import org.koin.core.context.startKoin
@@ -21,7 +25,14 @@ class DailyDishApplication :
       androidContext(this@DailyDishApplication)
       androidLogger(Level.INFO)
 
-      modules(module, dbModule)
+      modules(
+        module,
+        repositoryModule,
+        coreModule,
+        reminderModule
+      )
     }
+
+    get<ReminderService>().schedule()
   }
 }
