+++ app/src/main/java/io/krugosvet/dailydish/android/architecture/injection/Koin.kt
@@ -1,16 +1,14 @@
-@file:Suppress("RemoveExplicitTypeArguments")
-
 package io.krugosvet.dailydish.android.architecture.injection
 
 import androidx.fragment.app.Fragment
 import io.krugosvet.dailydish.android.architecture.view.GenericBaseActivity
-import io.krugosvet.dailydish.android.repository.Meal
-import io.krugosvet.dailydish.android.repository.MealRepository
 import io.krugosvet.dailydish.android.screen.addMeal.viewmodel.AddMealViewModel
 import io.krugosvet.dailydish.android.screen.container.view.ContainerActivity
 import io.krugosvet.dailydish.android.screen.mealList.view.MealVisualFactory
 import io.krugosvet.dailydish.android.screen.mealList.viewmodel.MealListViewModel
-import io.krugosvet.dailydish.android.service.*
+import io.krugosvet.dailydish.android.service.ImageService
+import io.krugosvet.dailydish.android.service.KeyboardService
+import io.krugosvet.dailydish.android.service.SnackbarService
 import org.koin.androidx.experimental.dsl.viewModel
 import org.koin.androidx.scope.lifecycleScope
 import org.koin.core.parameter.parametersOf
@@ -20,6 +18,7 @@ import org.koin.experimental.builder.single
 import kotlin.properties.ReadOnlyProperty
 import kotlin.reflect.KProperty
 
+@Suppress("RemoveExplicitTypeArguments")
 val module = module {
 
   scope<ContainerActivity> {
@@ -28,13 +27,7 @@ val module = module {
     scoped<KeyboardService>()
   }
 
-  single<DateService>()
-  single<ResourceService>()
-  single<IdGenerator>()
-
   single<MealVisualFactory>()
-  single<Meal.MealFactory>()
-  single<MealRepository>()
 
   viewModel<AddMealViewModel>()
   viewModel<MealListViewModel>()
