+++ app/src/main/java/io/krugosvet/dailydish/android/screen/addMeal/viewmodel/AddMealViewModel.kt
@@ -5,16 +5,16 @@ import androidx.lifecycle.map
 import androidx.lifecycle.viewModelScope
 import io.krugosvet.dailydish.android.architecture.extension.liveData
 import io.krugosvet.dailydish.android.architecture.viewmodel.ViewModel
-import io.krugosvet.dailydish.android.repository.Meal
-import io.krugosvet.dailydish.android.repository.MealRepository
+import io.krugosvet.dailydish.android.repository.meal.MealFactory
+import io.krugosvet.dailydish.android.repository.meal.MealRepository
 import io.krugosvet.dailydish.android.screen.addMeal.viewmodel.AddMealViewModel.Event
-import io.krugosvet.dailydish.android.service.DateService
+import io.krugosvet.dailydish.core.service.DateService
 import kotlinx.coroutines.launch
 
 class AddMealViewModel(
   private val mealRepository: MealRepository,
   private val dateService: DateService,
-  private val mealFactory: Meal.MealFactory
+  private val mealFactory: MealFactory
 ) :
   ViewModel<Event>() {
