--- app/src/main/java/io/krugosvet/dailydish/android/service/IdGenerator.kt
@@ -1,8 +1,0 @@
-package io.krugosvet.dailydish.android.service
-
-import java.util.*
-
-class IdGenerator {
-
-  fun generate(): Long = UUID.randomUUID().hashCode().toLong()
-}
