+++ core/build.gradle
@@ -15,16 +15,8 @@ android {
 }
 
 dependencies {
-    implementation project(":repository")
-    implementation project(":core")
-
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-
     implementation 'androidx.core:core-ktx:1.3.0'
-
-    // Jetpack Worker
-
-    implementation "androidx.work:work-runtime-ktx:$work_manager"
 
     // Koin
