+++ core/src/main/java/io/krugosvet/dailydish/core/service/IdGenerator.kt
@@ -1,0 +1,8 @@
+package io.krugosvet.dailydish.core.service
+
+import java.util.*
+
+class IdGenerator {
+
+  fun generate(): Long = UUID.randomUUID().hashCode().toLong()
+}
