+++ reminder/src/main/AndroidManifest.xml
@@ -1,4 +1,1 @@
-<manifest xmlns:android="http://schemas.android.com/apk/res/android"
-          package="io.krugosvet.dailydish.android.reminder">
-    /
-</manifest>
+<manifest package="io.krugosvet.dailydish.android.reminder" />
