+++ repository/build.gradle
@@ -1,6 +1,7 @@
 apply plugin: 'com.android.library'
 apply plugin: 'kotlin-android'
 apply plugin: 'kotlin-android-extensions'
+apply plugin: 'kotlin-kapt'
 
 android {
     compileSdkVersion 29
@@ -11,22 +12,28 @@ android {
         targetSdkVersion 29
         versionCode 1
         versionName "1.0"
+        consumerProguardFiles "consumer-rules.pro"
     }
 }
 
 dependencies {
-    implementation project(":repository")
     implementation project(":core")
 
     implementation "org.jetbrains.kotlin:kotlin-stdlib:$kotlin_version"
-
     implementation 'androidx.core:core-ktx:1.3.0'
 
-    // Jetpack Worker
+    // Jetpack Room
 
-    implementation "androidx.work:work-runtime-ktx:$work_manager"
+    implementation "androidx.room:room-runtime:$room_version"
+    implementation "androidx.room:room-ktx:$room_version"
+    kapt "androidx.room:room-compiler:$room_version"
 
     // Koin
 
     implementation "org.koin:koin-core:$koin_version"
+
+    // Coroutines
+
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-core:$coroutines_version"
+    implementation "org.jetbrains.kotlinx:kotlinx-coroutines-android:$coroutines_version"
 }
